Emacs specific settings
I am not putting this in the root because other emacs users might
consider it invasive.  Instead we check it into a subfolder that an
emacs user can instead symlink too.

This is similar to what we do with vscode, but instead we actually
check the files in for others to share.