Fully specify XVIZ 2 (minus declarative UI & treetable) (#66)
This provides a full schema for the XVIZ 2 protocol documented in the
main documentation.  There are some things to improve but that will
happen in further changes that update docs and spec at the same time.

Left out right now are the treetable and declarative UI JSON schema
specifications.  Since this is rather massive.