Fix schema tests and improve primitive examples (#103)
The schema test were using `test.only` so they could not be included
in the full test set, that is now fixed.  Coverage generation still
works.

The examples have been tweaked to allow for easier validation of
external XVIZ implementations.  The complex primitive example is a
super set of the simple one, and the object ID is a guid.  They also
now have examples of inline styles.