Move pose to explicit location in stream_set in spec & docs
A core part of XVIZ is knowing the location of the vehicle(s) so they
can be displayed relative to the other data.  This explicitly
specifies how to pass that data along and what format it comes in.

We are adding a `poses` map to the `stream_set` object, which already
contains all the core data like primitives and time series data. By
default if the `poses` field is present it must contain the default
pose stream `/vehicle_pose`.

Each field of that map contains the raw position and orientation of
the vehicle as well as a mapOrigin.

As seen in the discussions in #112 and #113.