Stream metadata coordinate and transform fields
In order to handle data from a variety of frames we can now specify
`coordinate` and `transform` fields stream.  The values for the
`coordinate` field are:

 * GEOGRAPHIC - no data, assumed long, lat
 * VEHICLE_RELATIVE - relative to pose, with optional transform
 * IDENTITY - default, no interpretation of the values
 * DYNAMIC - transform defines name of runtime transform function

The spec ensures that the transform field is only present when needed
for the VEHICLE_RELATIVE or DYNAMIC coordinate frames.

This is the specification part of #104 and #114