stream_set now maps stream to lists of variables (#144)
We want a single variable stream to be able to contain a unique
variable instance for a group of objects. Previous a variable stream
could only contain a single variable, mapping to a single object.

This lets us built more powerful UIs that show variables related to
many objects at the same time.