Fix Schema to handle lists of all primitive types
It was previously restricted to just `circle` and `point` types, now
the bug has been fixed.  The reason for the large number of files is
lists of primitives should be of a single kind only, so cannot make
one list of with a single primitive of each type. This is enables
more efficient and dense packing of primitives in the future.

We also added some negative examples to define what a our generic
value type cannot be.