Update schema and provided compatible proto3 IDL files
The goal here is to increase adoption by letting people build quick
clients with based on the protobuf IDL files and websocket
library. Note we are excluding declarative UI because that is meant as
a pure text protocol that is easy for humans to edit.

The downside is we have to content with protobuf not supported:

 - Maps of arrays
 - Polymorphic types
 - Variant fields
 - No typedefs

So the major changes we have to do for this is:

 - Move common fields from primitive, and annotation into base sub
   type
 - Replace polymorphic lists of primitives with lists per types
 - Add explicitly typed lists for time series and variables

This also means protobuf will only ever support a subset of the JSON
schema, with that subset being:

 - Numeric only colors (no hex or string support)
 - Flattened arrays only (no nested array support)

(Maybe we'll even have some gRPC endpoints someday as well).