Fix XVIZ v1 parsing and enable dynamic XVIZ version control
To enable log level XVIZ version support with the same app I have
split what was `xvizConfig.version` into `xvizConfig.supportedVersions`
and `xvizSettings.currentMajorVersion`

`supportedVersions` is an array that will throw if the determined
version is not in the list. Defaults to [1, 2].

`currentMajorVersion` is determined at metadata parse time and stored in
xvizSettings. Defaults to 1 which is required for teleassist until XVIZ_TASK
makes metadata available.

Misc fixes:
- prefer inline primitive.type over v2 which has type as parent key of
  primitives array.
- fix futures parsing and add test cases for v1 formatted data
- fix object_id assignment in primitive parsing