Define and update TreeTable schema (#191)
We have translated the documented schema into a formal JSON and
protobuf schema with a set of invalid and valid examples.

The schema has changed to having all column values encoded as strings
for now.  In the future when we can figure out a more efficient
protobuf compatible encoding we'll make a backwards compatible switch
to that.