Fully flesh out XVIZ stream metadata
We want to be able say exactly what data is on what stream so we can
inform the front end of the fact.  This properly sets up a two tiered
system where you define the category a field is in, like "primitive"
then say what exactly it is like "circle".

The one interesting choice here is that we use the "scalar_type" field
to represent the type of both "time_series" and "variable" data.