Specification for reconfigure message
This allows a client to adjust how a server transform the data it
sends in XVIZ format.

In order to make this happen I had to add support for the
google.protobuf.Struct well known type to protobuf.js.  The upside it
lets us basically embed JSON directly into protobuf for use cases like
configuration updates.  It also does it in a way that is faster or
more dense than raw text.