Unstable transform_point_in_time message spec & docs
The `transform_point_in_time` message gives a client the ability to
for a set of XVIZ streams at a particular point in time.  This is
useful for generate 3D or 2D images of logs, or continuously viewing a
time of interest as you reconfigure the XVIZ transform done by server.