Session protocol docs & transform message specs (#226)
This explicitly defines the XVIZ 2 session protocol and all the
messages it contains.  We describe in details the flow of:

 - connect
 - client sends `start`
 - server sends `metadata`
 - optional: client sends `reconfigure`
 - client sends `transform_log`
 - server sends `state_update` (many of them)
 - server sends `transform_log_done`
 - optional: server sends `error`

To fully specify this we defined the follow new message:

 - `reconfigure` (this was just the envelope support)
 - `transform_log`
 - `transform_log_done`
 - `error`

Also some links were fixed in the documentation page, and we fixed
the `start` message schema and added more examples.