Address an XVIZ v1 code path for a 'no data' entry
The distinction between 'absense of data' vs 'no data' is not being
handled in the parsing of XVIZ v1. Specifically when we transform from
XVIZ to an internal structure we are not handling this case correctly
result in an assertion error downstream from parsing in the
LogSynchronizer.

The fix applied here is to ensure when parsing v1 we handle the empty
stream array case and mark the entry with the appropriate timestamp
rather than returning the empty object.  This allows the downstream
logic to properly abort search when an entry without data is encoutered.

XVIZ Details:
Since XVIZ does not model a time to live, we rely on the presence of
a stream entry.  In the case of 'no data', which means we have a source
datum but there is no entry so we should treat the data as such and not
continue to search for data within the TIME_WINDOW.

An issue has been open up to investigate the current state of this case
for XVIZ 2 https://github.com/uber/xviz/issues/235