Update example server with live looping support (#288)
This updates the server to loop a small data set forever.  It's sort of an experiment to see how valuable a tool like this is.  At it's heart it does several things

 - Remove `log_info` `start_time` and `end_time` when in `--live` mode
 - Load all data into memory, needed for fast repacking when in `--loop` mode
 - Loop the playback index it gets to end of the dataset
 - Adjust timestamps in data and repack so that times arrow stays in one direction

Along the way I noticed that streetscape.gl does breaks when timing information is removed from the metadata, I think it might be a transformation issue.

This server now is neither a decent example, or good production grade implementation.  We need to move this into a properly factored package in the future.