xviztool and middleware framework (#290)
* Build schema and move JSON schema files

Actually build the code so it can be used in other tools, and
consolidate all the schema files in one easier to refer to schema
folder at `modules/schema/schema`. This matches the
`modules/schema/proto` folder.

* xviztool and middleware framework

The `xviztool` is a home for all the command line programs we need to
drive XVIZ development.  It's built upon a "middleware framework"
where you can pass a stack of objects that can analyze and act on the
incoming and outgoing streams of XVIZ messages and open/close events.

Right now we have a `dump` command (the "." is the log name, but our
test server only serves up one log) that shows you what a server is
sending:

    $ ./bin/xviztool dump ws://localhost:8081 . --metadata --oneline
    [CONNECTED]
    [< START]
    [> METADATA] version: 2.0.0
    [CONNECTION CLOSED]

And `validate` command to see if the data is correct (yeah we need to
fix up our examples):

    $ ./bin/xviztool validate ws://localhost:8081 . --metadata --condensed
    ┌──────────┬───────┬─────────┬───────┬───────────────┐
    │ Type     │ Count │ Invalid │ Inv % │ Unique Errors │
    ├──────────┼───────┼─────────┼───────┼───────────────┤
    │ METADATA │ 1     │ 1       │ 100.0 │ 1             │
    └──────────┴───────┴─────────┴───────┴───────────────┘

Lots of future work including an RFC and/or more official docs.  But
first on the list are the ability to act on files, and save data for
more analysis.