Update server to add in log_info as needed (#315)
Some logs contain metadata without the "log_info" section that
contains the end and start time of the log itself.  The example
server has this information so we add it in if needed.

In general we need to decide on how much metadata is mandatory
and state what clients should do when no metadata is present.