WIP: "xviz log" command
This allow you to save data from an XVIZ server to disk
so you can server it up with the example server.  It's waiting
on two things:
 - tests
 - binary support

Need to figure out how to to support async tests properly, probably
adding a callback parameter to the middleware stack for any async
operations?