Add package @xviz/io
@xviz/io consolidates general reading and writing of the
XVIZ specified data format.

Prior to this functions and classes that operate on XVIZ data
where split across @xviz/builder and @xviz/parser. They will
be moved into @xviz/io in subsequent changes.

Motivation for this is due to @xviz/parser restructuring data that
no longer can be validated against our specification and validation
tooling.  This made it difficult to modify XVIZ and validate it is
still conforming.  Given that, this module should operate only
at the XVIZ specification level.