Change XVIZFormatter into an actual 'writer' as XVIZFormatWriter
XVIZFormatWriter is just a convenient, but common, class to construct
the proper writer in proxy based on the desired format output.

As such it makes more sense to be an actually writer and just proxy
through.

XVIZFormatWriter differs in one very important way.  It takes XVIZData
and not objects (like current writers). Eventually the other writers
will support XVIZData also.