Change the XVIZ of 'Frame' to 'Message'
A 'Frame' is often overloaded, and in nearly all case
in XVIZ we mean an XVIZ Message.  We have created a distinction
where if we use 'Frame' it is not an individual message but the
set of data that constitutes a visual 'Frame' which could come from
multiple messages.  Therefore, 'Frame' is still used in @xviz/parse

- tighten up our flattening logic to exclude arrays with less than 20
  elements

[] todo: ensure backwards compat in anything not io or server
   specifically builder must support getFrame ()
[] review spec files
[] review parity with builder/spec & io/spec files (there is
   some duplication due to xvizWriter and xvizBinaryWriter etc