Lint protobuf files, move to xviz.v2 package, ensure bazel compat (#476)
We are using Uber's prototool 1.7 for lint with the provided
prototool.yaml setting which rules we are not following. These are not
being followed so that JSON emitted and consumed by the protobuf types
matches that in the JSON schema.

Bazel compatibility allows other projects to pull in our source
releases and use the protobuf rules directly.

See the `docs/protocol-formats/protobuf-protocol.md` for more details
on Bazel and lint/formatting.