Fix json formatting of an Array with TypedArray elements
The KITTI conversion creates point clouds as an Array of TypedArrays.
This was not working when formatted as json because the conversion only
hands the property name and made assumptions about the typed arrays
being the top-level.

As a result it was creating an addition array nesting that broke when
it encounted an Array of TypedArrays.

The fix tracks the array nesting and if encounters a nested array,
returns an array from the TypedArray w/o additional nesting

Fixed in both @xviz/builder & @xviz/io