Move xviz decoding to @xviz/io (#498)
XVIZ can be formatted as JSON, GLB and Protobuf and handling
of the format detection and decoding was originally in @xviz/parser.

The creation of @xviz/io was consolidate general formatting concerns
and to ease getting to the XVIZ data.  This change moves the details of
that detection solely within @xviz/io and removes the cross package
dependencies.  Now @xviz/io controls all the format handling in a single place.

- consolidate XVIZ format handling in @xviz/io
- remove @xviz/io dependency on @xviz/builder
- add parser webworker build changes to access xviz & schema by source
- Invert the draco dependency due to size, now it must be supplied directly by the user
- update @xviz/parser to use @xviz/io
- adjust build to ensure schema is built first due to dependencies in
   other modules on protobuf defintions