Protobuf changes for protobuf.js instances
Protobuf.js uses prototype object creation which has the effect of
looking up the prototype chain for properties even if the instance
does not have it set.

This requires the use of 'hasOwnProperty()' in some places to
distinguish between the instance and the prototype property.

The library also assigns default values for fields, such as the
empty array or a value of 0. Our parsing would look at presences
to determine if a property was meaningfully set, but now we need to
have additional validation due to the default values set by the
protobuf.js library.