Add order independent TreeTable Row Child creation (#602)
This tracks all row instances created and delays finalizing
the structure rather than doing it upon each row creation.

The recent python fix to properly calling "reset()" exposed a bug in
the behavior of the UIPrimitive builder. It treated creating a row
the same way a Primitive builder would when changing the geometry type.

Instead a "row" is part of the TreeTable so it should just be additive
and not cause a reset.