Pass primitive 'type' to allow per-type default styles
Normally all style have the same default value if they represent
the same attribute.  However a change was made to enable per-primitive
default style changes so make sure that is reflected in the interface.

In general, i'd prefer to expose the styling in a different way so it
could be customized.