Add workerfarm ID to enable separate capacity control for large messages (#637)
This adds a workerId field to parseXVIZMessage() that allows for separate worker pools.

Arbitrarily dropping an XVIZ Message in the default pool can
result in loosing a stream that contains the most recent data. However, in a situation where
you know what streams a message contains you can apply a custom settings to
the worker capacity.