Improve time_series handling and add XVIZObject attributes
time_series can vary on time & object_id, however prior to this
we where only assigning a single value to a parsed stream.

This change improves the runtime structure to store time_series
as an array internally after parsing.

With the above improvement, time_series entries related to an object can
now be accessed via the XVIZObject via object.getAttributes.

The attribute name will be the last portion of the stream name, if
following the '/' convention for path-like naming, else the whole stream
name.

Care in streetscape.gl must be taken to ensure compatibility with this
change.